# server/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateTeacherQuestionRequest" : {
        "properties" : {
          "classIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "from" : {
            "type" : "string"
          },
          "stops" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "title" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "to" : {
            "type" : "string"
          }
        },
        "required" : [ "title" ],
        "type" : "object"
      },
...
      "JsonNode" : { },
...
      "TeacherQuestion" : {
        "properties" : {
          "answeredChildren" : {
            "format" : "int32",
            "type" : "integer"
          },
          "classIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "firstTryAccuracy" : {
            "format" : "double",
            "type" : "number"
          },
          "from" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "schoolId" : {
            "type" : "string"
          },
          "sentAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "stops" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "teacherId" : {
            "type" : "string"
          },
          "teacherName" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "to" : {
            "type" : "string"
          },
          "totalChildren" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "TeacherQuestionChildResult" : {
        "properties" : {
          "answered" : {
            "format" : "int32",
            "type" : "integer"
          },
          "childId" : {
            "type" : "string"
          },
          "classId" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "stars" : {
            "format" : "int32",
            "type" : "integer"
          },
          "stops" : {
            "items" : {
              "$ref" : "#/components/schemas/TeacherQuestionStopResult"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "TeacherQuestionResults" : {
        "properties" : {
          "children" : {
            "items" : {
              "$ref" : "#/components/schemas/TeacherQuestionChildResult"
            },
            "type" : "array"
          },
          "firstTryAccuracy" : {
            "format" : "double",
            "type" : "number"
          },
          "questionId" : {
            "type" : "string"
          },
          "stopIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TeacherQuestionStopResult" : {
        "properties" : {
          "answered" : {
            "type" : "boolean"
          },
          "answeredAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "correct" : {
            "type" : "boolean"
          },
          "stars" : {
            "format" : "int32",
            "type" : "integer"
          },
          "stopId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateTeacherQuestionRequest" : {
        "properties" : {
          "classIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "from" : {
            "type" : "string"
          },
          "stops" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "title" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "to" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/children/{id}/attempts" : {
      "post" : {
        "operationId" : "uploadAttempts",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Children" ]
      }
    },
...
    "/children/{id}/teacher-questions/{questionId}" : {
      "get" : {
        "operationId" : "childTeacherQuestion",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "questionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Teacher questions" ]
      }
    },
    "/children/{id}/teacher-questions/{questionId}/answers" : {
      "post" : {
        "operationId" : "childTeacherAnswers",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "questionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Teacher questions" ]
      }
    },
...
    "/teacher/questions" : {
      "get" : {
        "operationId" : "teacherQuestions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeacherQuestion"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Teacher questions" ]
      },
      "post" : {
        "operationId" : "createTeacherQuestion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTeacherQuestionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeacherQuestion"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "Teacher questions" ]
      }
    },
    "/teacher/questions/{id}" : {
      "put" : {
        "operationId" : "updateTeacherQuestion",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTeacherQuestionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeacherQuestion"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Teacher questions" ]
      }
    },
    "/teacher/questions/{id}/results" : {
      "get" : {
        "operationId" : "teacherQuestionResults",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeacherQuestionResults"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Teacher questions" ]
      }
    },
    "/teacher/questions/{id}/send" : {
      "post" : {
        "operationId" : "sendTeacherQuestion",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeacherQuestion"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Teacher questions" ]
      }
    },
...
  },
...
}